Вывод сообщения на случай, если точек 0
