+++ src/const.js
@@ -1,0 +1,7 @@
+const MIN_NUMBER = 1;
+const MAX_NUMBER = 10;
+
+const DATE_FORMAT_DATE = 'DD MMM';
+const DATE_FORMAT_TIME = 'HH:mm';
+
+export { MIN_NUMBER, MAX_NUMBER, DATE_FORMAT_DATE, DATE_FORMAT_TIME };
