+++ src/presenter/trips-presenter.js
@@ -3,11 +3,8 @@ import TripEventListView from '../view/trip-event-list-view';
 import ListSortView from '../view/list-sort-view.js';
 import EditPointView from '../view/edit-point-view.js';
 import PointView from '../view/trip-point-view.js';
-import ListFilterView from '../view/list-filter-view.js';
-// import NewPointView from '../view/add-new-point-view.js';
 import { isEscapeKey } from '../util.js';
-
-const tripEventsElement = document.querySelector('.trip-events');
+import ListEmptyView from '../view/list-empty-view.js';
 
 export default class TripPresenter {
   #tripComponent = new TripEventListView();
@@ -23,14 +20,20 @@ export default class TripPresenter {
 
   init() {
     this.#listPoints = [...this.#pointsModel.points];
-    render(new ListFilterView(), this.#pointContainer );
-    render(new ListSortView(), tripEventsElement);
-    render(this.#tripComponent, tripEventsElement);
-    // render(new NewPointView(), this.#tripComponent.element);
-    for (let i = 0; i < this.#listPoints.length; i++) {
-      // render(new PointView({point: this.#listPoints[i]}), this.#tripComponent.element);
-      this.#renderPoint(this.#listPoints[i]);
+
+    this.#renderPointsList();
+  }
+
+  #renderPointsList() {
+    if (!this.#listPoints.length) {
+      render(new ListEmptyView(), this.#pointContainer);
+      return;
     }
+
+    render(new ListSortView(), this.#pointContainer);
+    render(this.#tripComponent, this.#pointContainer);
+    // render(new NewPointView(), this.#tripComponent.element);
+    this.#listPoints.forEach((point) => this.#renderPoint(point));
   }
 
   #renderPoint(point){
@@ -50,8 +53,6 @@ export default class TripPresenter {
 
     const replaceFormToPoint = () => {
       this.#tripComponent.element.replaceChild(pointComponent.element, pointEditComponent.element);
-      editRollupBtn.removeEventListener('click', onCloseEditForm);
-      editPointForm.removeEventListener('submit', onCloseEditForm);
       document.removeEventListener('keydown', onEscKeyDown);
     };
